+++ build.sh
@@ -8,7 +8,7 @@ PROJECT="Form.xcodeproj"
 SCHEME="Form"
 
 IOS_SDK="${IOS_SDK:-"iphonesimulator13.0"}"
-IOS_DESTINATION_PHONE="${IOS_DESTINATION_PHONE:-"OS=13.0,name=iPhone Xs"}"
+IOS_DESTINATION_PHONE="${IOS_DESTINATION_PHONE:-"OS=13.0,name=iPhone 11"}"
 
 usage() {
 cat << EOF